separate database seeder and reader from main
