+++ Makefile
@@ -1,4 +1,5 @@
 # Simple makefile
+# @author William Tabi <c11342646@mydit.ie>
 
 CC=g++
 LD=gcc
@@ -6,20 +7,28 @@ LD=gcc
 # Define object files and debug info
 CCFLAGS=-c -g
 
-OBJS=main.o Mysql.o
-
 # Linker info: Include libraries and debug info
 LDFLAGS=-I/usr/include/mysql -lmysqlclient -lm
 
-main: $(OBJS)
-	@echo "Linking"
+create: create.o Mysql.o
+	@echo "Linking Database seeder"
 	# @see http://ubuntuforums.org/showthread.php?t=1666018&p=10351570#post10351570
 	# @see http://stackoverflow.com/questions/6045809/problem-with-g-and-undefined-reference-to-gxx-personality-v0?answertab=votes#tab-top
-	$(LD) $(LDFLAGS) $(OBJS) -o main `mysql_config --cflags --libs` -lstdc++
+	$(LD) $(LDFLAGS) create.o Mysql.o -o create `mysql_config --cflags --libs` -lstdc++
 
-main.o: main.cpp
-	@echo "Compiling main!"
-	$(CC) $(LDFLAGS) $(CCFLAGS) main.cpp -o main.o
+read: read.o Mysql.o
+	@echo "Linking Reader"
+	# @see http://ubuntuforums.org/showthread.php?t=1666018&p=10351570#post10351570
+	# @see http://stackoverflow.com/questions/6045809/problem-with-g-and-undefined-reference-to-gxx-personality-v0?answertab=votes#tab-top
+	$(LD) $(LDFLAGS) read.o Mysql.o -o read `mysql_config --cflags --libs` -lstdc++
+
+create.o: create.cpp
+	@echo "Compiling database seed!"
+	$(CC) $(LDFLAGS) $(CCFLAGS) create.cpp -o create.o
+
+read.o: read.cpp
+	@echo "Compiling database reader!"
+	$(CC) $(LDFLAGS) $(CCFLAGS) read.cpp -o read.o
 
 Mysql.o: Mysql.cpp
 	@echo "Compiling mysql class"
@@ -28,4 +37,5 @@ Mysql.o: Mysql.cpp
 clean:
 	@echo "Removing object files and executables"
 	rm *.o
-	rm main
+	rm create
+	rm read
